src/compliance/policy_engine.py
"""Policy rule engine for custom compliance checking."""

import yaml
from pathlib import Path
from typing import Any, Dict, List

from src.core.logger import get_logger
from src.core.models import ComplianceResult

logger = get_logger(__name__)


class PolicyEngine:
    """Custom policy rule engine."""

    def __init__(self, policy_file: Path):
        """Initialize the policy engine."""
        self.logger = get_logger(self.__class__.__name__)
        self.policy_file = policy_file
        self.policies = self._load_policies()

    def _load_policies(self) -> Dict[str, Any]:
        """Load policies from YAML file."""
        if not self.policy_file.exists():
            self.logger.warning("Policy file not found", file=str(self.policy_file))
            return {}

        try:
            with open(self.policy_file, "r") as f:
                return yaml.safe_load(f) or {}
        except Exception as e:
            self.logger.error("Failed to load policies", error=str(e))
            return {}

    def evaluate_policies(self) -> List[ComplianceResult]:
        """
        Evaluate all policies.

        Returns:
            List of compliance results
        """
        self.logger.info("Evaluating custom policies")
        results = []

        policies = self.policies.get("policies", [])
        for policy in policies:
            try:
                passed = self._evaluate_policy(policy)
                result = ComplianceResult(
                    framework="custom",
                    rule_id=policy.get("id", "unknown"),
                    rule_name=policy.get("name", "Unknown policy"),
                    passed=1 if passed else 0,
                    description=policy.get("description", ""),
                    remediation=policy.get("remediation", ""),
                )
                results.append(result)
            except Exception as e:
                self.logger.error("Error evaluating policy", policy_id=policy.get("id"), error=str(e))
                result = ComplianceResult(
                    framework="custom",
                    rule_id=policy.get("id", "unknown"),
                    rule_name=policy.get("name", "Unknown policy"),
                    passed=0,
                    description=f"Error evaluating policy: {str(e)}",
                )
                results.append(result)

        return results

    def _evaluate_policy(self, policy: Dict[str, Any]) -> bool:
        """Evaluate a single policy rule."""
        # This is a placeholder for policy evaluation logic
        # In a full implementation, this would parse and execute policy rules
        # For now, we'll return a simple check
        condition = policy.get("condition", {})
        check_type = condition.get("type")

        # Simple condition evaluation
        if check_type == "file_exists":
            path = Path(condition.get("path", ""))
            return path.exists()
        elif check_type == "file_permission":
            path = Path(condition.get("path", ""))
            if not path.exists():
                return False
            expected_perms = condition.get("permissions", "644")
            # Simplified permission check
            return True  # Would need proper permission checking
        else:
            return False

