requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
click==8.1.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Data processing
pyyaml==6.0.1
pandas==2.1.3

# Reporting
jinja2==3.1.2
reportlab==4.0.7

# Logging & monitoring
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.11.0
ruff==0.1.6
mypy==1.7.1

